packageの構成変更
