+++ src/main/scala/com/github/imomushi8/systra/entity/Position.scala
@@ -1,7 +1,7 @@
-package com.github.imomushi8.systra.util
+package com.github.imomushi8.systra.entity
 
 /*--------------------------------------------------------------------------------------------*/
-import com.github.imomushi8.systra.{ID, Side, Price, Size, TimeStamp}
+import com.github.imomushi8.systra.util.{ID, Side, Price, Size, TimeStamp}
 /*--------------------------------------------------------------------------------------------*/
