+++ src/test/scala/BackTestContractSpec.scala
@@ -1,7 +1,7 @@
 import com.github.imomushi8.systra._
 import com.github.imomushi8.systra.backtest._
-import com.github.imomushi8.systra.backtest.{BackTest, BTMarket}
 import com.github.imomushi8.systra.util._
+import com.github.imomushi8.systra.entity._
 import com.github.imomushi8.systra.report.PositionReport
 
 import org.scalacheck.{Arbitrary, Gen}
